Move common SqlScriptingContextManager initialization code from SqlScriptingInterpreterSuite
