+++ sql/core/src/test/scala/org/apache/spark/sql/scripting/SqlScriptingInterpreterSuite.scala
@@ -19,7 +19,7 @@ package org.apache.spark.sql.scripting
 
 import org.apache.spark.{SparkConf, SparkException, SparkNumberFormatException}
 import org.apache.spark.sql.{AnalysisException, QueryTest, Row}
-import org.apache.spark.sql.catalyst.{QueryPlanningTracker, SqlScriptingContextManager}
+import org.apache.spark.sql.catalyst.QueryPlanningTracker
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.plans.logical.CompoundBody
 import org.apache.spark.sql.classic.{DataFrame, Dataset}
@@ -33,7 +33,10 @@ import org.apache.spark.sql.test.SharedSparkSession
  * Output from the interpreter (iterator over executable statements) is then checked - statements
  *   are executed and output DataFrames are compared with expected outputs.
  */
-class SqlScriptingInterpreterSuite extends QueryTest with SharedSparkSession {
+class SqlScriptingInterpreterSuite
+    extends QueryTest
+    with SharedSparkSession
+    with SqlScriptingTestUtils {
 
   // Tests setup
   override protected def sparkConf: SparkConf = {
@@ -44,19 +47,9 @@ class SqlScriptingInterpreterSuite extends QueryTest with SharedSparkSession {
   private def runSqlScript(
       sqlText: String,
       args: Map[String, Expression] = Map.empty): Array[DataFrame] = {
-    val interpreter = SqlScriptingInterpreter(spark)
     val compoundBody = spark.sessionState.sqlParser.parsePlan(sqlText).asInstanceOf[CompoundBody]
 
-    // Initialize context so scopes can be entered correctly.
-    val context = new SqlScriptingExecutionContext()
-    val executionPlan = interpreter.buildExecutionPlan(compoundBody, args, context)
-    context.frames.append(new SqlScriptingExecutionFrame(
-      executionPlan, SqlScriptingFrameType.SQL_SCRIPT))
-    executionPlan.enterScope()
-
-    val handle =
-      SqlScriptingContextManager.create(new SqlScriptingContextManagerImpl(context))
-    handle.runWith {
+    withSqlScriptingContextManager(spark, compoundBody, args) { executionPlan =>
       executionPlan.getTreeIterator.flatMap {
         case statement: SingleStatementExec =>
           if (statement.isExecuted) {
